refactor(parsers): move stateless section parsers out of CanOpenReaderBase
Moves the stateless [DeviceInfo]/[Comments]/[SupportedModules]/
[DynamicChannels]/[Tools] parsers into the new internal static
CanOpenSectionParsers class and drops the thin ParseSectionsFrom* wrappers
in favor of direct IniParser calls. This removes the entire CA1822
suppression cluster (11 attributes) and leaves CanOpenReaderBase with only
the members that vary per format (KnownSectionNames, ParseFileInfo,
ParseObjectDictionary/ParseObject/ParseSubObjects/ParseSubObject,
IsKnownSection and shared static helpers). No behavior change.

Closes #340
